requirements.txt
pip~=25.1
Django~=5.1.4
psycopg2-binary~=2.9.10
djangorestframework~=3.16.0
python-dotenv~=1.1.0